pyproject.toml
[project]
name = "ai-syncfiles"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.11.0",
    "google>=3.0.0",
    "google-auth>=2.43.0",
    "langchain>=1.1.0",
    "langchain-classic>=1.0.0",
    "langchain-ollama>=1.0.0",
    "langchain-openai>=1.1.0",
    "ollama>=0.6.1",
    "python-dotenv>=1.2.1",
]
